Fixed timer bug in level select
testing fix for bug with ui resolution

+++ Assets/Scripts/LevelSelect.cs
@@ -54,9 +54,9 @@ public class LevelSelect : MonoBehaviour
 
         Button button = buttonObject.GetComponentInChildren<Button>();
         TextMeshProUGUI buttonText = buttonObject.GetComponentInChildren<TextMeshProUGUI>();
-
+        
         string minutes = ((int)levelData.bestTime / 60).ToString("00");
-        string seconds = (levelData.bestTime % 60).ToString("00");
+        string seconds = ((int)levelData.bestTime % 60).ToString("00");
         string milliseconds = ((int)((levelData.bestTime * 1000) % 1000)).ToString("000");
 
         buttonText.text = levelName + "\nBest Time: " + minutes + ":" + seconds + ":" + milliseconds;
